requirements.txt
streamlit
mido
pygame
requests
Pillow
